Renaming Epinano_Variants.dev.py to Epinano_variants.py in the main folder. README updated.
