+++ README.md
@@ -160,33 +160,27 @@ Below is a simple introduction of programs' usage information.
 **Epinano_Variants**, outputs a feature table *sample.per.site.var.csv*, which contains base-calling ‘error’ information for each reference position. Please note that by default, the feature table *sample.per_site.5mer.csv* that was generated by default in EpiNano 1.1 (which contains the same base-called features organized in 5-mer windows) is not generated by default any more. If you want to generate this file, please use the script Slide_Variants.py. 
 ```  
 python Epinano_Variants.py -h
-usage: Epinano_Variants.py [-h] [-R REFERENCE] [-b BAM] [-s SAM2TSV]
-                           [-n NUMBER_CPUS] [-T TYPE]
+usage: Epinano_Variants.py [-h] [-r REFERENCE] [-b BAM]
+                           [-c NUMBER_CPUS]
 
 optional arguments:
   -h, --help            show this help message and exit
-  -n NUMBER_CPUS, --number_cpus NUMBER_CPUS
+  -c NUMBER_CPUS, --number_cpus NUMBER_CPUS
                         number of CPUs
-  -T TYPE, --type TYPE  reference types, which is either g(enome) or
-                        t(ranscriptome);
 
 Required Arguments:
-  -R REFERENCE, --reference REFERENCE
+  -r REFERENCE, --reference REFERENCE
                         reference file indexed with samtools faidx and with
                         sequence dictionary created using picard
                         CreateSequenceDictionary
   -b BAM, --bam BAM     bam file; if given; no need to offer reads file;
                         mapping will be skipped
-  -s SAM2TSV, --sam2tsv SAM2TSV
-                        /path/to/sam2tsv.jar; needed unless a sam2tsv.jar
-                        produced file is already given
  
 ```  
-The '--type' flag indicates the type of mapping, i.e., mapping to reference genome (*g*) versus mapping to referece transcriptome (*t*). If the reads were mapped to a genome reference with splicing mapping, '--type g' should be specified, and EpiNano will distinguish and process separately the reads mapped to forward and reverse strands. Otherwise, by default, the script assumes the bam file was generated by mapping the reads to reference transcriptome and the reads were only mapped to the forward strand. 
 
 Example: 
 ```  
-python $EPINANO_HOME/Epinano_Variants.py -n 6 -R reference.fasta -b sample.reads.bam -s  $EPINANO_HOME/misc/sam2tsv.jar --type t 
+python $EPINANO_HOME/Epinano_Variants.py -c 6 -r reference.fasta -b sample.reads.bam 
 ```
 Note 1: it is possible to organize the variants in any kmer length using *$EPINANO_HOME/misc/Slide_Variants.py*.
